requirements.txt
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.23.0
scipy>=1.9.0
Pillow>=9.0.0
tqdm>=4.64.0
git+https://github.com/openai/CLIP.git
